requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
MarkupSafe==3.0.2
motor==3.7.0
numpy==2.2.3
pandas==2.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
